requirements.txt
alembic==1.2.1
astroid==2.2.5
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
mccabe==0.6.1
Pillow==6.2.0
psycopg2==2.8.4
pylint==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
